Move Env to correct location
