+++ std/jvm/src/main/scala/cats/effect/std/EnvCompanionPlatform.scala
@@ -17,7 +17,7 @@
 package cats.effect.std
 
 import cats.effect.kernel.Sync
-import cats.effect.std.syntax.Env
+import cats.effect.std.Env
 
 import scala.collection.immutable.Map
